+++ LuckyoneUI/Locales/ruRU.lua
@@ -213,6 +213,7 @@ L["Outlaw"] = "Головорез"
 L["Paladin"] = "Паладин"
 L["People found out some characters have a big amount of hidden quests which will cause performance issues. This button will untrack all your quests, including the hidden ones and might give you an increase in average FPS."] = "Игроки обнаружили, что у некоторых персонажей есть большое количество скрытых заданий, что может вызывать проблемы с производительностью. Эта кнопка отключит отслеживание всех Ваших заданий, включая скрытые, и может повысить средний FPS."
 L["Performance"] = "Производительность"
+L["Performance"] = true
 L["PermoksAccountManager profile has been set."] = "Профиль PermoksAccountManager установлен."
 L["Plater profile has been set."] = "Профиль Plater установлен."
 L["Please click the button below to apply Luckyones Nameplates profile."] = "Пожалуйста, нажмите кнопку ниже, чтобы применить профиль индикаторов здоровья Luckyone."